Cache thumbnails in development.
Thumbnails are saved to a cache directory after being created for the first time so that `Image#resize!` becomes a no-op on subsequent requests.

For cache invalidation, the filename of a cached thumbnail includes a SHA2 checksum of the original image. (The checksum is calculated on each request, but this is still noteably faster than resizing the original image, especially for larger image sizes.)

The cache directory defaults to 'tmp/simple-thumbnailer-cache' and can be customized via the `cache_dir` option.

There is no automatic removal of stale cached thumbnails. (Just manually remove the whole cache directory from time to time to get rid of old files.)